Move create_cont_describe function to tables module.
