+++ tableone/tables.py
@@ -1,9 +1,11 @@
 from typing import Optional
+import warnings
 
 import numpy as np
 import pandas as pd
 
 from tableone.statistics import Statistics
+from tableone.validators import InputError
 
 
 class Tables:
@@ -299,3 +301,86 @@ class Tables:
             df_cat = df_cat.swaplevel(0, 1, axis=1).sort_index(axis=1, level=0)
 
         return df_cat
+
+    def create_cont_describe(self,
+                             data: pd.DataFrame,
+                             ddof,
+                             t1_summary,
+                             dip_test,
+                             tukey_test,
+                             normal_test,
+                             continuous,
+                             groupby: Optional[str] = None) -> pd.DataFrame:
+        """
+        Describe the continuous data.
+
+        Parameters
+        ----------
+            data : pandas DataFrame
+                The input dataset.
+
+        Returns
+        ----------
+            df_cont : pandas DataFrame
+                Summarise the continuous variables.
+        """
+        # wrapper for std with ddof
+        def std(x):
+            return self.statistics._std(x, ddof)
+
+        aggfuncs = ['count', 'mean', 'median', std,
+                    self.statistics._q25, self.statistics._q75,
+                    'min', 'max', t1_summary]
+
+        if dip_test:
+            aggfuncs.append(self.statistics._hartigan_dip)
+
+        if tukey_test:
+            aggfuncs.append(self.statistics._outliers)
+            aggfuncs.append(self.statistics._far_outliers)
+
+        if normal_test:
+            aggfuncs.append(self.statistics._normality)
+
+        # coerce continuous data to numeric
+        cont_data = data[continuous].apply(pd.to_numeric, errors='coerce')
+        # check all data in each continuous column is numeric
+        bad_cols = cont_data.count() != data[continuous].count()
+        bad_cols = cont_data.columns[bad_cols]
+        if len(bad_cols) > 0:
+            msg = ("The following continuous column(s) have "
+                   "non-numeric values: {variables}. Either specify the "
+                   "column(s) as categorical or remove the "
+                   "non-numeric values.").format(variables=bad_cols.values)
+            raise InputError(msg)
+
+        # check for coerced column containing all NaN to warn user
+        for column in cont_data.columns[cont_data.count() == 0]:
+            self._non_continuous_warning(column)
+
+        if groupby:
+            # add the groupby column back
+            cont_data = cont_data.merge(data[[groupby]], left_index=True, right_index=True)
+
+            # group and aggregate data
+            df_cont = pd.pivot_table(cont_data, columns=[groupby], aggfunc=aggfuncs)
+        else:
+            # if no groupby, just add single group column
+            df_cont = cont_data.apply(aggfuncs).T  # type: ignore
+            df_cont.columns.name = 'Overall'
+            df_cont.columns = pd.MultiIndex.from_product([df_cont.columns, ['Overall']])
+
+        df_cont.index = df_cont.index.rename('variable')
+
+        # remove prefix underscore from column names (e.g. _std -> std)
+        agg_rename = df_cont.columns.levels[0]  # type: ignore
+        agg_rename = [x[1:] if x[0] == '_' else x for x in agg_rename]
+        df_cont.columns = df_cont.columns.set_levels(agg_rename, level=0)  # type: ignore
+
+        return df_cont
+
+    # warnings
+    def _non_continuous_warning(self, c):
+        msg = ("'{}' has all non-numeric values. Consider including "
+               "it in the list of categorical variables.").format(c)
+        warnings.warn(msg, RuntimeWarning, stacklevel=2)
